Fixed the problem with babelrc interfering wiht the build
