--- babel.config.js
@@ -1,13 +1,0 @@
-// Only used in tests
-module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
-      },
-    ],
-  ],
-};
