--- docs/demo.382444c8.css
@@ -1,2 +1,0 @@
-body{font-family:Roboto,sans-serif;margin:0;padding:0}section{background-attachment:fixed}section,section a{color:#fff}@media (max-width:570px){#editorNode{margin-left:-10px;margin-right:-10px;width:auto;align-self:stretch}}@media (max-width:1400px){section{padding:20px;display:flex;flex-direction:column;align-items:center}section>*{width:100%;max-width:580px}section>:last-child{margin-top:20px}#editorNode>table{min-width:100%}}@media (min-width:1400px){section{padding:80px;display:flex;min-height:80vh;align-items:center;justify-content:center;font-size:20px}section>*{box-sizing:border-box;max-width:600px;margin:0 40px;flex-grow:1;flex-shrink:1}}ul.checks{list-style:none;padding:0}ul.checks li{margin-bottom:10px}ul.checks li:before{content:"\2611";margin-right:10px}h1,h2{font-size:44px;font-weight:600}#editorNode{overflow:auto}pre>code{background:#2e3548;color:#e6f7fd;border-radius:4px;font-size:14px;max-width:100%}pre.auto,pre>code{display:block;padding:20px;overflow:auto}pre.auto{white-space:pre;background:#fff;color:#000;font-family:monospace;line-height:20px;margin-bottom:40px;font-size:12px;max-height:80vh}section.sample{color:rgba(0,0,0,.8)}section.sample>:first-child{order:2}
-/*# sourceMappingURL=https://renanlecaro.github.io/importabular/demo.382444c8.css.map */
