--- docs/demo.c971d610.js
@@ -1,8 +1,0 @@
-parcelRequire=function(e,r,t,n){var i,o="function"==typeof parcelRequire&&parcelRequire,u="function"==typeof require&&require;function f(t,n){if(!r[t]){if(!e[t]){var i="function"==typeof parcelRequire&&parcelRequire;if(!n&&i)return i(t,!0);if(o)return o(t,!0);if(u&&"string"==typeof t)return u(t);var c=new Error("Cannot find module '"+t+"'");throw c.code="MODULE_NOT_FOUND",c}p.resolve=function(r){return e[t][1][r]||r},p.cache={};var l=r[t]=new f.Module(t);e[t][0].call(l.exports,p,l,l.exports,this)}return r[t].exports;function p(e){return f(p.resolve(e))}}f.isParcelRequire=!0,f.Module=function(e){this.id=e,this.bundle=f,this.exports={}},f.modules=e,f.cache=r,f.parent=o,f.register=function(r,t){e[r]=[function(e,r){r.exports=t},{}]};for(var c=0;c<t.length;c++)try{f(t[c])}catch(e){i||(i=e)}if(t.length){var l=f(t[t.length-1]);"object"==typeof exports&&"undefined"!=typeof module?module.exports=l:"function"==typeof define&&define.amd?define(function(){return l}):n&&(this[n]=l)}if(parcelRequire=f,i)throw i;return f}({"Bzn5":[function(require,module,exports) {
-"use strict";function t(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function e(t,e){for(var n=0;n<e.length;n++){var i=e[n];i.enumerable=i.enumerable||!1,i.configurable=!0,"value"in i&&(i.writable=!0),Object.defineProperty(t,i.key,i)}}function n(t,n,i){return n&&e(t.prototype,n),i&&e(t,i),t}function i(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var r=function(){function e(n){var r=this,a=n.data,h=void 0===a?[]:a,u=n.node,l=void 0===u?null:u,c=n.onChange,d=void 0===c?null:c,f=n.minRows,v=void 0===f?1:f,y=n.maxRows,m=void 0===y?1/0:y,p=n.minCols,x=void 0===p?1:p,g=n.maxCols,b=void 0===g?1/0:g,k=n.css,w=void 0===k?"":k,C=n.width,E=void 0===C?"100%":C,D=n.height,R=void 0===D?"80vh":D;if(t(this,e),i(this,"t",null),i(this,"_",1),i(this,"p",1),i(this,"g",new o),i(this,"T",["mousedown","mouseenter","mouseup","mouseleave","touchstart","touchend","touchmove","keydown","paste","cut","copy"]),i(this,"paste",function(t){if(!r.R){t.preventDefault();for(var e=function(t){try{var e=(t.clipboardData||window.clipboardData).getData("text/html"),n=document.createElement("iframe");document.body.appendChild(n),n.contentWindow.document.open(),n.contentWindow.document.write(e),n.contentWindow.document.close();var i=n.contentWindow.document.querySelectorAll("tr"),r=[];if(Array.prototype.forEach.call(i,function(t,e){var n=t.querySelectorAll("td");Array.prototype.forEach.call(n,function(t,n){var i=t.innerText;r[e]||(r[e]=[]),r[e][n]=i})}),document.body.removeChild(n),r.length)return r}catch(t){}return(t.clipboardData||window.clipboardData).getData("text").split(/\r\n|\n|\r/).map(function(t){return t.split("")})}(t),n=r.O,i=n.rx,o=n.ry,a={x:i[0],y:o[0]},s=0;s<e.length;s++)for(var h=0;h<e[0].length;h++)r.j(a.x+h,a.y+s,e[s][h]);r.U(function(){r.$=a,r.B={x:a.x+e[0].length-1,y:a.y+e.length-1}}),r.v()}}),i(this,"copy",function(t){if(!r.R){var e=r.H();e&&(t.preventDefault(),t.clipboardData.setData("text/html",function(t){var e=document.createElement("table");return e.setAttribute("lang",navigator.language),t.forEach(function(t){var n=document.createElement("tr");e.appendChild(n),t.forEach(function(t){var e=document.createElement("td");n.appendChild(e),e.innerText=t})}),'<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN"><html lang="'.concat(navigator.language,'"><head>\n<meta http-equiv="content-type" content="text/html; charset=utf-8"/><title></title></head><body>').concat(e.outerHTML,"</body></html>")}(e)),t.clipboardData.setData("text/plain",e.map(function(t){return t.join("")}).join("\n")))}}),i(this,"cut",function(t){r.R||(r.copy(t),r.P(""))}),i(this,"keydown",function(t){t.ctrlKey||r.$&&("Escape"===t.key&&r.R&&(t.preventDefault(),r.W(),r.N()),"Enter"===t.key&&(t.preventDefault(),r.q(!1,t.shiftKey?-1:1)),"Tab"===t.key&&(t.preventDefault(),r.q(!0,t.shiftKey?-1:1)),r.R||("Delete"!==t.key&&"Backspace"!==t.key||(t.preventDefault(),r.P("")),"ArrowDown"===t.key&&(t.preventDefault(),r.F({y:1},t.shiftKey)),"ArrowUp"===t.key&&(t.preventDefault(),r.F({y:-1},t.shiftKey)),"ArrowLeft"===t.key&&(t.preventDefault(),r.F({x:-1},t.shiftKey)),"ArrowRight"===t.key&&(t.preventDefault(),r.F({x:1},t.shiftKey))),1!==t.key.length||r.R||r.U(function(){var t=r.Y,e=t.x,n=t.y;r.G({x:e,y:n}),r.J(e,n).firstChild.value=""}))}),i(this,"st",!1),i(this,"$",null),i(this,"B",null),i(this,"O",{rx:[0,0],ry:[0,0]}),i(this,"R",null),i(this,"Y",null),i(this,"mousedown",function(t){if(!r.mobile){if(3===t.which&&!r.R&&r.tt()){var e=new Range,n=r.O,i=n.rx,o=n.ry;return e.setStart(r.J(i[0],o[0]),0),e.setEnd(r.J(i[0],o[0]),1),r.cwd.getSelection().removeAllRanges(),void r.cwd.getSelection().addRange(e)}r.U(function(){r.tbody.style.userSelect="none",r.B=r.$=r.Y=r.it(t),r.st=!0})}}),i(this,"mouseenter",function(t){r.mobile||r.st&&r.U(function(){r.B=r.it(t)})}),i(this,"ht",null),i(this,"et",null),i(this,"mouseup",function(t){r.mobile||3!==t.which&&r.st&&r.U(function(){r.B=r.it(t),r.nt(),r.ht&&r.ht>Date.now()-300&&r.et.x===r.B.x&&r.et.y===r.B.y&&r.G(r.B),r.ht=Date.now(),r.et=r.B})}),i(this,"mouseleave",function(t){t.target===r.tbody&&r.st&&r.nt()}),i(this,"touchstart",function(t){r.R||(r.mobile=!0,r.moved=!1)}),i(this,"touchend",function(t){r.mobile&&(r.R||r.moved||(r.U(function(){r.B=r.$=r.Y=r.it(t)}),r.G(r.Y)))}),i(this,"touchmove",function(t){r.mobile&&(r.moved=!0)}),i(this,"N",function(){if(r.R){var t=r.R,e=t.x,n=t.y,i=r.J(e,n);i.style.width="",i.style.height="";var o=i.firstChild;o.removeEventListener("blur",r.N),o.removeEventListener("keydown",r.rt),r.j(e,n,o.value),i.removeChild(o),r.R=null,r.k(i,e,n),r.v()}}),i(this,"rt",function(t){13===t.keyCode&&(r.N(),t.preventDefault())}),i(this,"S",function(t){var e=t.x,n=t.y;r.J(e,n).className=r.lt(e,n)}),i(this,"X",function(t){var e=t.x,n=t.y,i=r.J(e,n).firstChild;"DIV"===i.tagName&&(i.innerText=r.A(e,n)),r.S({x:e,y:n})}),!l)return console.error("Please call the constructor like this : new Importabular({node: document.body})");this.t=l,this.i={onChange:d,minRows:v,maxRows:m,minCols:x,maxCols:b,css:s+w},this.h={width:E,height:R,border:"none",background:"transparent"},this.o(),this.l(h),this.u({x:this.i.minCols-1,y:this.i.minRows-1}),this.m()}return n(e,[{key:"C",value:function(t){var e=t.x,n=t.y;return e>=0&&e<this.i.maxCols&&n>=0&&n<this.i.maxRows}},{key:"m",value:function(){var t=Math.ceil(this.iframe.contentWindow.innerHeight/40),e=Math.ceil(this.iframe.contentWindow.innerWidth/100);this.u({x:e-1,y:t-1})}},{key:"v",value:function(){this.i.onChange&&this.i.onChange(this.g.D())}},{key:"k",value:function(t,e,n){var i=document.createElement("div");t.setAttribute("x",e.toString()),t.setAttribute("y",n.toString());var r=this.A(e,n);r?i.innerText=r:i.innerHTML="&nbsp;",t.appendChild(i),this.S({x:e,y:n})}},{key:"o",value:function(){var t=this,e=document.createElement("iframe");this.iframe=e,this.t.appendChild(e);var n=e.contentWindow.document;this.cwd=n,n.open(),n.write('<html lang="'.concat(navigator.language,'"><body><style>').concat(this.i.css,"</style></body></html>")),n.close(),Object.assign(e.style,this.h);var i=document.createElement("table"),r=document.createElement("tbody");i.appendChild(r),n.body.appendChild(i),this.tbody=r,this.table=i;for(var o=0;o<this.p;o++){var a=document.createElement("tr");r.appendChild(a);for(var s=0;s<this._;s++)this.I(a,s,o)}this.T.forEach(function(e){return n.addEventListener(e,t[e],!0)})}},{key:"destroy",value:function(){var t=this;this.M(),this.T.forEach(function(e){return t.cwd.removeEventListener(e,t[e],!0)}),this.iframe.parentElement.removeChild(this.iframe)}},{key:"I",value:function(t,e,n){var i=document.createElement("td");t.appendChild(i),this.k(i,e,n)}},{key:"V",value:function(){if(!this.C({x:0,y:this.p}))return!1;this.p++;var t=this.p-1,e=document.createElement("tr");this.tbody.appendChild(e);for(var n=0;n<this._;n++)this.I(e,n,t);return!0}},{key:"L",value:function(){var t=this;if(!this.C({x:this._,y:0}))return!1;this._++;var e=this._-1;return Array.prototype.forEach.call(this.tbody.children,function(n,i){t.I(n,e,i)}),!0}},{key:"u",value:function(t){for(var e=t.x,n=t.y;e>this._-1&&this.L(););for(;n>this.p-1&&this.V(););}},{key:"H",value:function(){var t=this.O,e=t.rx,n=t.ry;if(e[0]===e[1])return null;for(var i=e[1]-e[0],r=n[1]-n[0],o=[],a=0;a<r;a++){o.push([]);for(var s=0;s<i;s++)o[a].push(this.A(e[0]+s,n[0]+a))}return o}},{key:"P",value:function(t){var e=this;this.K(this.O,function(n){var i=n.x,r=n.y;return e.j(i,r,t)}),this.K(this.O,this.X),this.v()}},{key:"F",value:function(t,e){var n=this,i=t.x,r=void 0===i?0:i,o=t.y,a=void 0===o?0:o,s=e?this.B:this.$,h={x:s.x+r,y:s.y+a};this.C(h)&&(this.N(),this.u(h),this.U(function(){e?n.B=h:n.$=n.B=n.Y=h}),this.Z(h))}},{key:"q",value:function(t){var e,n=this,i=arguments.length>1&&void 0!==arguments[1]?arguments[1]:1,r=this.Y||{x:0,y:0},o=r.x,s=r.y,h=this.tt(),u=h>1?this.O:{rx:[0,this.i.maxCols],ry:[0,this.i.maxRows]},l=u.rx,c=u.ry;if(t)e=a(o,s,i,l[0],l[1]-1,c[0],c[1]-1);else{var d=a(s,o,i,c[0],c[1]-1,l[0],l[1]-1);e={x:d.y,y:d.x}}this.C(e)&&(this.N(),this.u(e),this.U(function(){n.Y=e,h<=1&&(n.$=n.B=e)}),this.Z(e))}},{key:"Z",value:function(t){var e=t.x,n=t.y;this.J(e,n).scrollIntoView({behavior:"smooth",block:"nearest"})}},{key:"nt",value:function(){this.st=!1,this.tbody.style.userSelect=""}},{key:"G",value:function(t){var e=t.x,n=t.y;this.R={x:e,y:n};var i=this.J(e,n),r=i.getBoundingClientRect(),o=i.firstChild.getBoundingClientRect();Object.assign(i.style,{width:r.width-2,height:r.height}),i.removeChild(i.firstChild);var a=document.createElement("input");a.type="text",a.value=this.A(e,n),Object.assign(a.style,{width:o.width,height:o.height}),i.appendChild(a),a.focus(),a.addEventListener("blur",this.N),a.addEventListener("keydown",this.rt)}},{key:"M",value:function(){if(this.R){var t=this.R,e=t.x,n=t.y,i=this.J(e,n).firstChild;i.removeEventListener("blur",this.N),i.removeEventListener("keydown",this.rt)}}},{key:"W",value:function(){if(this.R){var t=this.R,e=t.x,n=t.y;this.J(e,n).firstChild.value=this.A(e,n)}}},{key:"U",value:function(t){var e=this.O;t(),this.O=this.ot(),this.K(e,this.S),this.K(this.O,this.S)}},{key:"ot",value:function(){if(!this.$)return{rx:[0,0],ry:[0,0]};var t=[this.$.x,this.B.x];t[0]>t[1]&&t.reverse();var e=[this.$.y,this.B.y];return e[0]>e[1]&&e.reverse(),{rx:[t[0],t[1]+1],ry:[e[0],e[1]+1]}}},{key:"K",value:function(t,e){for(var n=t.rx,i=t.ry,r=n[0];r<n[1];r++)for(var o=i[0];o<i[1];o++)this.C({x:r,y:o})&&e({x:r,y:o})}},{key:"tt",value:function(){var t=this.O,e=t.rx,n=t.ry;return(e[1]-e[0])*(n[1]-n[0])}},{key:"lt",value:function(t,e){var n=this.O,i=n.rx,r=n.ry,o="";return t>=i[0]&&t<i[1]&&e>=r[0]&&e<r[1]&&(o+=" selected",this.tt()>1&&(o+=" multi")),this.Y&&this.Y.x===t&&this.Y.y===e&&(o+=" focus"),this.R&&t===this.R.x&&e===this.R.y&&(o+=" editing"),o}},{key:"it",value:function(t){for(var e=t.target;!e.getAttribute("x")&&e.parentElement;)e=e.parentElement;return{x:parseInt(e.getAttribute("x"))||0,y:parseInt(e.getAttribute("y"))||0}}},{key:"setData",value:function(t){this.g.ct(),this.l(t);for(var e=0;e<this._;e++)for(var n=0;n<this.p;n++)this.X({x:e,y:n})}},{key:"getData",value:function(){return this.g.D()}},{key:"l",value:function(t){var e=this;t.forEach(function(t,n){t.forEach(function(t,i){e.j(i,n,t)})})}},{key:"j",value:function(t,e,n){this.C({x:t,y:e})&&(this.g.j(t,e,n),this.u({x:t+1,y:e+1}),this.X({x:t,y:e}))}},{key:"A",value:function(t,e){return this.g.A(t,e)}},{key:"J",value:function(t,e){return this.tbody.children[e].children[t]}}]),e}();exports.default=r;var o=function(){function e(){t(this,e),i(this,"g",{})}return n(e,[{key:"j",value:function(t,e,n){var i,r=this.g,o=0===n?"0":n?n.toString():"";o?(r[t]||(r[t]={}),r[t][e]=o):r[t]&&r[t][e]&&(delete r[t][e],i=r[t],0===Object.keys(i).length&&delete r[t])}},{key:"ct",value:function(){this.g={}}},{key:"A",value:function(t,e){var n=this.g;return n&&n[t]&&n[t][e]||""}},{key:"D",value:function(){var t=1,e=1;for(var n in this.g)for(var i in this.g[n])e=Math.max(e,parseInt(i)+1),t=Math.max(t,parseInt(n)+1);for(var r=[],o=0;o<e;o++){r.push([]);for(var a=0;a<t;a++)r[o].push(this.A(a,o))}return r}}]),e}();function a(t,e,n,i,r,o,a){if((t+=n)<i){if(r===1/0)return{x:i,y:e};if(t=r,--e<o){if(a===1/0)return{x:i,y:o};e=a}}return t>r&&(t=i,++e>a&&(e=o,t=i)),{x:t,y:e}}var s="\nhtml{\n  -ms-overflow-style: none;\n  scrollbar-width: none;\n}\n::-webkit-scrollbar {\n  visibility: hidden;\n}\n*{\n  box-sizing: border-box;\n}\nbody{\n  padding: 0; \n  margin: 0;\n}\ntable{\n  border-spacing: 0;\n  background: white;\n  border: 1px solid #ddd;\n  border-width: 0 1px 1px 0;\n  font-size: 16px;\n  font-family: sans-serif;\n  border-collapse: separate;\n}\ntd{\n  padding:0;\n  border: 1px solid;\n  border-color: #ddd transparent transparent #ddd; \n}\ntd.selected.multi:not(.editing){\n  background:#d7f2f9;\n} \ntd.focus:not(.editing){\n  border-color: black;\n} \ntd>*{\n  border:none;\n  padding:10px;\n  min-width:100px;\n  min-height: 40px;\n  font:inherit;\n  line-height: 20px;\n  color:inherit;\n  white-space: normal;\n}\ntd>div::selection {\n    color: none;\n    background: none;\n}\n";
-},{}],"n4SH":[function(require,module,exports) {
-
-},{}],"lgAh":[function(require,module,exports) {
-"use strict";var e=t(require("../dist/index.js"));function t(e){return e&&e.__esModule?e:{default:e}}require("./demo.css"),window.sheet=new e.default({data:[["Product ID","Product name","Price","Unit","Category"],["256","Sample product","2.5","Piece","Stuffs"],["122","Pre mix drink","5","Bottle","Drinks"]],node:document.getElementById("editorNode"),maxCols:5}),console.info("\n\nYou can play around with the instance on the page, try :\n\nsheet.getData()\nsheet.setData([['Hello','World']])\nsheet.destroy()\n\n"),Array.prototype.forEach.call(document.getElementsByClassName("readable"),function(e){var t=document.createElement("pre");t.className="auto";var n=e.innerText.trim().replace('from "./dist/index.js"','from "importabular"').replace('from "./src/index.js"','from "importabular/src/index.js"'),r=n.split("\n"),a=r[1].match(/ +/)[0].length;n=r.map(function(e,t){return t?e.slice(0,a).trim()?e.trim():e.slice(a):e}).join("\n"),t.innerText=n,e.parentElement.appendChild(t)});
-},{"../dist/index.js":"Bzn5","./demo.css":"n4SH"}]},{},["lgAh"], null)
-//# sourceMappingURL=https://renanlecaro.github.io/importabular/demo.c971d610.js.map
