+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "importabular",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "5kb javascript spreadsheet, let your users import their data from excel.",
   "main": "dist/index.js",
   "files": [
@@ -8,10 +8,10 @@
     "src/index.js"
   ],
   "scripts": {
-    "start": "parcel demo/index.html --out-dir docs",
+    "start": "npm run build-lib && rm -rf  docs/* && parcel demo/index.html --out-dir docs",
     "build-jsdoc": "documentation readme src/index.js --section=API",
     "build-doc": "rm -rf docs && parcel build demo/index.html --out-dir docs --public-url https://renanlecaro.github.io/importabular/ ",
-    "build-lib": "rm -rf dist/* && terser src/index.js --module --compress --mangle --module --keep-classnames --mangle-props regex=/^_/ --ecma 2015 > dist/index.js && gzip -k dist/index.js && ls -l dist",
+    "build-lib": "sh scripts/build.sh",
     "build": "npm run pretty && npm run build-doc && npm run build-lib && npm run build-jsdoc",
     "pretty": "prettier --write demo src",
     "test": "jest"
@@ -48,7 +48,9 @@
   },
   "homepage": "https://renanlecaro.github.io/importabular/",
   "devDependencies": {
+    "@babel/cli": "^7.11.6",
     "@babel/core": "^7.11.6",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/preset-env": "^7.11.5",
     "babel-jest": "^26.3.0",
     "documentation": "*",
