+++ src/_arrToHTML.test.js
@@ -1,5 +1,11 @@
-import {_arrToHTML} from "./_arrToHTML";
+import { _arrToHTML } from "./index";
 
-test('_arrToHTML',()=>{
-  expect(_arrToHTML([['A1']])).toContain('A1')
-})
+test("textContent behavior", () => {
+  const div = document.createElement("div");
+  div.textContent = "Hello";
+  expect(div.innerHTML).toEqual("Hello");
+});
+
+test("_arrToHTML", () => {
+  expect(_arrToHTML([["A1"]])).toContain("A1");
+});
