+++ src/_parsePasteEvent.test.js
@@ -1,21 +1,21 @@
-import {_parsePasteEvent} from "./_parsePasteEvent";
-import {_arrToHTML} from "./_arrToHTML";
+import { _parsePasteEvent } from "./index";
+import { _arrToHTML } from "./index";
 
-describe('_parsePasteEvent',()=>{
-  const baseTable=[
-    ['A1','A2','A3'],
-    ['B1','B2','B3']
-  ]
+test("_parsePasteEvent", () => {
+  const baseTable = [
+    ["A1", "A2", "A3"],
+    ["B1", "B2", "B3"],
+  ];
   const event = {
-    clipboardData:{
-      getData(type){
-        if(type==='text/html'){
-          return _arrToHTML(baseTable)
-        }else{
-          return baseTable.map(line=>line.join('\t')).join('\n')
+    clipboardData: {
+      getData(type) {
+        if (type === "text/html") {
+          return _arrToHTML(baseTable);
+        } else {
+          return baseTable.map((line) => line.join("\t")).join("\n");
         }
-      }
-    }
-  }
-  expect(_parsePasteEvent(event)).toEqual(baseTable)
-})
+      },
+    },
+  };
+  expect(_parsePasteEvent(event)).toEqual(baseTable);
+});
